refactor: run_lightcone simpler
