+++ src/py21cmfast/drivers/coeval.py
@@ -1,21 +1,464 @@
 """Compute simulations that evolve over redshift."""
 
+import contextlib
+import h5py
 import logging
 import numpy as np
 import os
+from hashlib import md5
 from pathlib import Path
-from typing import Any
+from typing import Any, Sequence
 
+from .. import __version__
+from .. import utils as _ut
+from .._cfg import config
 from ..c_21cmfast import lib
-from ..photoncons import _get_photon_nonconservation_data, setup_photon_cons
 from ..wrapper.globals import global_params
 from ..wrapper.inputs import AstroParams, CosmoParams, FlagOptions, UserParams
-from ..wrapper.outputs import Coeval, InitialConditions, PerturbedField
+from ..wrapper.outputs import (
+    BrightnessTemp,
+    HaloBox,
+    InitialConditions,
+    IonizedBox,
+    PerturbedField,
+    TsBox,
+    _OutputStruct,
+)
+from ..wrapper.photoncons import _get_photon_nonconservation_data, setup_photon_cons
 from . import single_field as sf
-from .param_config import InputParameters, _get_config_options, _setup_inputs
+from .param_config import InputParameters, _get_config_options
 from .single_field import set_globals
 
 logger = logging.getLogger(__name__)
+
+
+class _HighLevelOutput:
+    def get_cached_data(
+        self, kind: str, redshift: float, load_data: bool = False
+    ) -> _OutputStruct:
+        """
+        Return an OutputStruct object which was cached in creating this Coeval box.
+
+        Parameters
+        ----------
+        kind
+            The kind of object: "init", "perturb", "spin_temp", "ionize" or "brightness"
+        redshift
+            The (approximate) redshift of the object to return.
+        load_data
+            Whether to actually read the field data of the object in (call ``obj.read()``
+            after this function to do this manually)
+
+        Returns
+        -------
+        output
+            The output struct object.
+        """
+        if self.cache_files is None:
+            raise AttributeError(
+                "No cache files were associated with this Coeval object."
+            )
+
+        # TODO: also check this file, because it may have been "gather"d.
+
+        if kind not in self.cache_files:
+            raise ValueError(
+                f"{kind} is not a valid kind for the cache. Valid options: "
+                f"{self.cache_files.keys()}"
+            )
+
+        files = self.cache_files.get(kind, {})
+        # files is a list of tuples of (redshift, filename)
+
+        redshifts = np.array([f[0] for f in files])
+
+        indx = np.argmin(np.abs(redshifts - redshift))
+        fname = files[indx][1]
+
+        if not os.path.exists(fname):
+            raise OSError(
+                "The cached file you requested does not exist (maybe it was removed?)."
+            )
+
+        kinds = {
+            "init": InitialConditions,
+            "perturb_field": PerturbedField,
+            "halobox": HaloBox,
+            "ionized_box": IonizedBox,
+            "spin_temp": TsBox,
+            "brightness_temp": BrightnessTemp,
+        }
+        cls = kinds[kind]
+
+        return cls.from_file(fname, load_data=load_data)
+
+    def gather(
+        self,
+        fname: str | None | Path = None,
+        kinds: Sequence | None = None,
+        clean: bool | dict = False,
+        direc: str | Path | None = None,
+    ) -> Path:
+        """Gather the cached data associated with this object into its file."""
+        kinds = kinds or [
+            "init",
+            "perturb_field",
+            "halobox",
+            "ionized_box",
+            "spin_temp",
+            "brightness_temp",
+        ]
+
+        clean = kinds if clean and not hasattr(clean, "__len__") else clean or []
+        if any(c not in kinds for c in clean):
+            raise ValueError(
+                "You are trying to clean cached items that you will not be gathering."
+            )
+
+        direc = Path(direc or config["direc"]).expanduser().absolute()
+        fname = Path(fname or self.get_unique_filename()).expanduser()
+
+        if not fname.exists():
+            fname = direc / fname
+
+        for kind in kinds:
+            redshifts = (f[0] for f in self.cache_files[kind])
+            for i, z in enumerate(redshifts):
+                cache_fname = self.cache_files[kind][i][1]
+
+                obj = self.get_cached_data(kind, redshift=z, load_data=True)
+                with h5py.File(fname, "a") as fl:
+                    cache = (
+                        fl.create_group("cache") if "cache" not in fl else fl["cache"]
+                    )
+                    kind_group = (
+                        cache.create_group(kind) if kind not in cache else cache[kind]
+                    )
+
+                    zstr = f"z{z:.2f}"
+                    if zstr not in kind_group:
+                        z_group = kind_group.create_group(zstr)
+                    else:
+                        z_group = kind_group[zstr]
+
+                    obj.write_data_to_hdf5_group(z_group)
+
+                    if kind in clean:
+                        os.remove(cache_fname)
+        return fname
+
+    def _get_prefix(self):
+        return "{name}_z{redshift:.4}_{{hash}}_r{seed}.h5".format(
+            name=self.__class__.__name__,
+            redshift=float(self.redshift),
+            seed=self.random_seed,
+        )
+
+    def _input_rep(self):
+        return "".join(
+            repr(getattr(self, inp))
+            for inp in [
+                "user_params",
+                "cosmo_params",
+                "astro_params",
+                "flag_options",
+                "global_params",
+            ]
+        )
+
+    def get_unique_filename(self):
+        """Generate a unique hash filename for this instance."""
+        return self._get_prefix().format(
+            hash=md5((self._input_rep() + self._particular_rep()).encode()).hexdigest()
+        )
+
+    def _write(self, direc=None, fname=None, clobber=False):
+        """
+        Write the lightcone to file in standard HDF5 format.
+
+        This method is primarily meant for the automatic caching. Its default
+        filename is a hash generated based on the input data, and the directory is
+        the configured caching directory.
+
+        Parameters
+        ----------
+        direc : str, optional
+            The directory into which to write the file. Default is the configuration
+            directory.
+        fname : str, optional
+            The filename to write, default a unique name produced by the inputs.
+        clobber : bool, optional
+            Whether to overwrite existing file.
+
+        Returns
+        -------
+        fname : str
+            The absolute path to which the file was written.
+        """
+        direc = os.path.expanduser(direc or config["direc"])
+
+        if fname is None:
+            fname = self.get_unique_filename()
+
+        if not os.path.isabs(fname):
+            fname = os.path.abspath(os.path.join(direc, fname))
+
+        if not clobber and os.path.exists(fname):
+            raise FileExistsError(
+                f"The file {fname} already exists. If you want to overwrite, set clobber=True."
+            )
+
+        with h5py.File(fname, "w") as f:
+            # Save input parameters as attributes
+            for k in [
+                "user_params",
+                "cosmo_params",
+                "flag_options",
+                "astro_params",
+                "global_params",
+            ]:
+                q = getattr(self, k)
+                kfile = "_globals" if k == "global_params" else k
+                grp = f.create_group(kfile)
+
+                try:
+                    dct = q.self
+                except AttributeError:
+                    dct = q
+
+                for kk, v in dct.items():
+                    if v is None:
+                        continue
+                    with contextlib.suppress(TypeError):
+                        grp.attrs[kk] = v
+            if self.photon_nonconservation_data is not None:
+                photon_data = f.create_group("photon_nonconservation_data")
+                for k, val in self.photon_nonconservation_data.items():
+                    photon_data[k] = val
+
+            f.attrs["redshift"] = self.redshift
+            f.attrs["random_seed"] = self.random_seed
+            f.attrs["version"] = __version__
+
+        self._write_particulars(fname)
+
+        return fname
+
+    def _write_particulars(self, fname):
+        pass
+
+    def save(self, fname=None, direc=".", clobber: bool = False):
+        """Save to disk.
+
+        This function has defaults that make it easy to save a unique box to
+        the current directory.
+
+        Parameters
+        ----------
+        fname : str, optional
+            The filename to write, default a unique name produced by the inputs.
+        direc : str, optional
+            The directory into which to write the file. Default is the current directory.
+
+        Returns
+        -------
+        str :
+            The filename to which the box was written.
+        """
+        return self._write(direc=direc, fname=fname, clobber=clobber)
+
+    @classmethod
+    def _read_inputs(cls, fname):
+        kwargs = {}
+        with h5py.File(fname, "r") as fl:
+            glbls = dict(fl["_globals"].attrs)
+            kwargs["redshift"] = fl.attrs["redshift"]
+
+            if "photon_nonconservation_data" in fl.keys():
+                d = fl["photon_nonconservation_data"]
+                kwargs["photon_nonconservation_data"] = {k: d[k][...] for k in d.keys()}
+
+        return kwargs, glbls
+
+    @classmethod
+    def read(cls, fname, direc="."):
+        """Read a lightcone file from disk, creating a LightCone object.
+
+        Parameters
+        ----------
+        fname : str
+            The filename path. Can be absolute or relative.
+        direc : str
+            If fname, is relative, the directory in which to find the file. By default,
+            both the current directory and default cache and the  will be searched, in
+            that order.
+
+        Returns
+        -------
+        LightCone :
+            A :class:`LightCone` instance created from the file's data.
+        """
+        if not os.path.isabs(fname):
+            fname = os.path.abspath(os.path.join(direc, fname))
+
+        if not os.path.exists(fname):
+            raise FileExistsError(f"The file {fname} does not exist!")
+
+        park, glbls = cls._read_inputs(fname)
+        boxk = cls._read_particular(fname)
+
+        with global_params.use(**glbls):
+            out = cls(**park, **boxk)
+
+        return out
+
+
+class Coeval(_HighLevelOutput):
+    """A full coeval box with all associated data."""
+
+    def __init__(
+        self,
+        redshift: float,
+        initial_conditions: InitialConditions,
+        perturbed_field: PerturbedField,
+        ionized_box: IonizedBox,
+        brightness_temp: BrightnessTemp,
+        ts_box: TsBox | None = None,
+        halobox: HaloBox | None = None,
+        cache_files: dict | None = None,
+        photon_nonconservation_data=None,
+        _globals=None,
+    ):
+
+        # Check that all the fields have the same input parameters.
+        InputParameters.from_output_structs(
+            (
+                initial_conditions,
+                perturbed_field,
+                halobox,
+                ionized_box,
+                brightness_temp,
+                ts_box,
+            )
+        )
+
+        self.redshift = redshift
+        self.init_struct = initial_conditions
+        self.perturb_struct = perturbed_field
+        self.ionization_struct = ionized_box
+        self.brightness_temp_struct = brightness_temp
+        self.halobox_struct = halobox
+        self.spin_temp_struct = ts_box
+
+        self.cache_files = cache_files
+
+        self.photon_nonconservation_data = photon_nonconservation_data
+        # A *copy* of the current global parameters.
+        self.global_params = _globals or dict(global_params.items())
+
+        # Expose all the fields of the structs to the surface of the Coeval object
+        for box in [
+            initial_conditions,
+            perturbed_field,
+            halobox,
+            ionized_box,
+            brightness_temp,
+            ts_box,
+        ]:
+            if box is None:
+                continue
+            for field in box._get_box_structures():
+                setattr(self, field, getattr(box, field))
+
+        # For backwards compatibility
+        if hasattr(self, "velocity_z"):
+            self.velocity = self.velocity_z
+
+    @classmethod
+    def get_fields(cls, spin_temp: bool = True, hbox: bool = True) -> list[str]:
+        """Obtain a list of name of simulation boxes saved in the Coeval object."""
+        pointer_fields = []
+        for cls in [InitialConditions, PerturbedField, IonizedBox, BrightnessTemp]:
+            pointer_fields += cls.get_pointer_fields()
+
+        if spin_temp:
+            pointer_fields += TsBox.get_pointer_fields()
+
+        if hbox:
+            pointer_fields += HaloBox.get_pointer_fields()
+
+        return pointer_fields
+
+    @property
+    def user_params(self):
+        """User params shared by all datasets."""
+        return self.brightness_temp_struct.user_params
+
+    @property
+    def cosmo_params(self):
+        """Cosmo params shared by all datasets."""
+        return self.brightness_temp_struct.cosmo_params
+
+    @property
+    def flag_options(self):
+        """Flag Options shared by all datasets."""
+        return self.brightness_temp_struct.flag_options
+
+    @property
+    def astro_params(self):
+        """Astro params shared by all datasets."""
+        return self.brightness_temp_struct.astro_params
+
+    @property
+    def random_seed(self):
+        """Random seed shared by all datasets."""
+        return self.brightness_temp_struct.random_seed
+
+    def _particular_rep(self):
+        return ""
+
+    def _write_particulars(self, fname):
+        for name in ["init", "perturb", "ionization", "brightness_temp", "spin_temp"]:
+            struct = getattr(self, f"{name}_struct")
+            if struct is not None:
+                struct.write(fname=fname, write_inputs=False)
+
+                # Also write any other inputs to any of the constituent boxes
+                # to the overarching attrs.
+                with h5py.File(fname, "a") as fl:
+                    for inp in struct._inputs:
+                        if inp not in fl.attrs and inp not in [
+                            "user_params",
+                            "cosmo_params",
+                            "flag_options",
+                            "astro_params",
+                            "global_params",
+                        ]:
+                            fl.attrs[inp] = getattr(struct, inp)
+
+    @classmethod
+    def _read_particular(cls, fname):
+        kwargs = {}
+
+        with h5py.File(fname, "r") as fl:
+            for output_class in _ut.OutputStruct._implementations():
+                if output_class.__name__ in fl:
+                    kwargs[_ut.camel_to_snake(output_class.__name__)] = (
+                        output_class.from_file(fname)
+                    )
+
+        return kwargs
+
+    def __eq__(self, other):
+        """Determine if this is equal to another object."""
+        return (
+            isinstance(other, self.__class__)
+            and other.redshift == self.redshift
+            and self.user_params == other.user_params
+            and self.cosmo_params == other.cosmo_params
+            and self.flag_options == other.flag_options
+            and self.astro_params == other.astro_params
+        )
 
 
 def get_logspaced_redshifts(min_redshift: float, z_step_factor: float, zmax: float):
@@ -144,13 +587,10 @@ def run_coeval(
 
     # We can go ahead and purge some of the stuff in the init_box, but only if
     # it is cached -- otherwise we could be losing information.
-    try:
+    with contextlib.suppress(OSError):
         initial_conditions.prepare_for_perturb(
             flag_options=flag_options, force=always_purge
         )
-    except OSError:
-        pass
-
     if perturbed_field:
         if redshift is not None and any(
             p.redshift != z for p, z in zip(perturbed_field, redshift)
@@ -193,23 +633,17 @@ def run_coeval(
         )
 
         if user_params.MINIMIZE_MEMORY:
-            try:
+            with contextlib.suppress(OSError):
                 p.purge(force=always_purge)
-            except OSError:
-                pass
-
         perturb_.append(p)
 
     perturbed_field = perturb_
 
     # Now we can purge init_box further.
-    try:
+    with contextlib.suppress(OSError):
         initial_conditions.prepare_for_halos(
             flag_options=flag_options, force=always_purge
         )
-    except OSError:
-        pass
-
     # get the halos (reverse redshift order)
     pt_halos = []
     if inputs.flag_options.USE_HALO_FIELD and not inputs.flag_options.FIXED_HALO_GRIDS:
@@ -219,23 +653,18 @@ def run_coeval(
             pt_halos += [sf.perturb_halo_list(redshift=z, halo_field=halos, **kw)]
 
             # we never want to store every halofield
-            try:
+            with contextlib.suppress(OSError):
                 pt_halos[i].purge(force=always_purge)
-            except OSError:
-                pass
             halos_desc = halos
 
         # reverse to get the right redshift order
         pt_halos = pt_halos[::-1]
 
     # Now we can purge init_box further.
-    try:
+    with contextlib.suppress(OSError):
         initial_conditions.prepare_for_spin_temp(
             flag_options=flag_options, force=always_purge
         )
-    except OSError:
-        pass
-
     if (
         flag_options.PHOTON_CONS_TYPE == 1
         and np.amin(redshifts) < global_params.PhotonConsEndCalibz
@@ -326,20 +755,14 @@ def run_coeval(
         )
 
         if pf is not None:
-            try:
+            with contextlib.suppress(OSError):
                 pf.purge(force=always_purge)
-            except OSError:
-                pass
-
         if ph2 is not None:
-            try:
+            with contextlib.suppress(OSError):
                 ph2.purge(force=always_purge)
-            except OSError:
-                pass
-
         # we only need the SFR fields at previous redshifts for XraySourceBox
         if hb is not None:
-            try:
+            with contextlib.suppress(OSError):
                 hb.prepare(
                     keep=[
                         "halo_sfr",
@@ -349,9 +772,6 @@ def run_coeval(
                     ],
                     force=always_purge,
                 )
-            except OSError:
-                pass
-
         if z in redshift:
             logger.debug(f"PID={os.getpid()} doing brightness temp for z={z}")
             ib_tracker[redshift.index(z)] = ib2
