+++ src/py21cmfast/wrapper/photoncons.py
@@ -37,7 +37,6 @@ from .wrapper.inputs import (
     FlagOptions,
     UserParams,
     global_params,
-    validate_all_inputs,
 )
 
 logger = logging.getLogger(__name__)
